Move some helpers from XamlHelpers to ActionHelpers
