+++ source/uwp/Renderer/lib/AdaptiveMediaRenderer.cpp
@@ -3,6 +3,8 @@
 #include "pch.h"
 
 #include "AdaptiveMedia.h"
+
+#include "ActionHelpers.h"
 #include "AdaptiveMediaRenderer.h"
 #include "AdaptiveElementParserRegistration.h"
 #include "MediaHelpers.h"
@@ -60,7 +62,7 @@ namespace AdaptiveNamespace
         CreatePosterContainerWithPlayButton(posterImage.Get(), renderContext, renderArgs, &posterContainer);
 
         ComPtr<IUIElement> touchTargetUIElement;
-        XamlHelpers::WrapInTouchTarget(adaptiveCardElement, posterContainer.Get(), nullptr, renderContext, true, L"Adaptive.SelectAction", &touchTargetUIElement);
+        ActionHelpers::WrapInTouchTarget(adaptiveCardElement, posterContainer.Get(), nullptr, renderContext, true, L"Adaptive.SelectAction", &touchTargetUIElement);
 
         // Create a panel to hold the poster and the media element
         ComPtr<IStackPanel> mediaStackPanel =
