+++ source/uwp/Renderer/lib/XamlHelpers.cpp
@@ -7,7 +7,6 @@
 #include "ActionHelpers.h"
 #include "ElementTagContent.h"
 #include "TileControl.h"
-#include "AdaptiveShowCardActionRenderer.h"
 #include "AdaptiveImage.h"
 
 using namespace Microsoft::WRL;
@@ -82,117 +81,6 @@ namespace AdaptiveNamespace::XamlHelpers
         }
 
         return S_OK;
-    }
-
-    void WrapInTouchTarget(_In_ IAdaptiveCardElement* adaptiveCardElement,
-                           _In_ IUIElement* elementToWrap,
-                           _In_ IAdaptiveActionElement* action,
-                           _In_ IAdaptiveRenderContext* renderContext,
-                           bool fullWidth,
-                           const std::wstring& style,
-                           _COM_Outptr_ IUIElement** finalElement)
-    {
-        ComPtr<IAdaptiveHostConfig> hostConfig;
-        THROW_IF_FAILED(renderContext->get_HostConfig(&hostConfig));
-
-        if (ActionHelpers::WarnForInlineShowCard(renderContext, action, L"Inline ShowCard not supported for SelectAction"))
-        {
-            // Was inline show card, so don't wrap the element and just return
-            ComPtr<IUIElement> localElementToWrap(elementToWrap);
-            localElementToWrap.CopyTo(finalElement);
-            return;
-        }
-
-        ComPtr<IButton> button =
-            XamlHelpers::CreateXamlClass<IButton>(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_Button));
-
-        ComPtr<IContentControl> buttonAsContentControl;
-        THROW_IF_FAILED(button.As(&buttonAsContentControl));
-        THROW_IF_FAILED(buttonAsContentControl->put_Content(elementToWrap));
-
-        ComPtr<IAdaptiveSpacingConfig> spacingConfig;
-        THROW_IF_FAILED(hostConfig->get_Spacing(&spacingConfig));
-
-        UINT32 cardPadding = 0;
-        if (fullWidth)
-        {
-            THROW_IF_FAILED(spacingConfig->get_Padding(&cardPadding));
-        }
-
-        ComPtr<IFrameworkElement> buttonAsFrameworkElement;
-        THROW_IF_FAILED(button.As(&buttonAsFrameworkElement));
-
-        // We want the hit target to equally split the vertical space above and below the current item.
-        // However, all we know is the spacing of the current item, which only applies to the spacing above.
-        // We don't know what the spacing of the NEXT element will be, so we can't calculate the correct spacing
-        // below. For now, we'll simply assume the bottom spacing is the same as the top. NOTE: Only apply spacings
-        // (padding, margin) for adaptive card elements to avoid adding spacings to card-level selectAction.
-        if (adaptiveCardElement != nullptr)
-        {
-            ABI::AdaptiveNamespace::Spacing elementSpacing;
-            THROW_IF_FAILED(adaptiveCardElement->get_Spacing(&elementSpacing));
-            UINT spacingSize;
-            THROW_IF_FAILED(GetSpacingSizeFromSpacing(hostConfig.Get(), elementSpacing, &spacingSize));
-            double topBottomPadding = spacingSize / 2.0;
-
-            // For button padding, we apply the cardPadding and topBottomPadding (and then we negate these in the margin)
-            ComPtr<IControl> buttonAsControl;
-            THROW_IF_FAILED(button.As(&buttonAsControl));
-            THROW_IF_FAILED(buttonAsControl->put_Padding({(double)cardPadding, topBottomPadding, (double)cardPadding, topBottomPadding}));
-
-            double negativeCardMargin = cardPadding * -1.0;
-            double negativeTopBottomMargin = topBottomPadding * -1.0;
-
-            THROW_IF_FAILED(buttonAsFrameworkElement->put_Margin(
-                {negativeCardMargin, negativeTopBottomMargin, negativeCardMargin, negativeTopBottomMargin}));
-        }
-
-        // Style the hit target button
-        THROW_IF_FAILED(
-            XamlHelpers::SetStyleFromResourceDictionary(renderContext, style.c_str(), buttonAsFrameworkElement.Get()));
-
-        if (action != nullptr)
-        {
-            // If we have an action, use the title for the AutomationProperties.Name
-            HString title;
-            THROW_IF_FAILED(action->get_Title(title.GetAddressOf()));
-
-            ComPtr<IDependencyObject> buttonAsDependencyObject;
-            THROW_IF_FAILED(button.As(&buttonAsDependencyObject));
-
-            ComPtr<IAutomationPropertiesStatics> automationPropertiesStatics;
-            THROW_IF_FAILED(
-                GetActivationFactory(HStringReference(RuntimeClass_Windows_UI_Xaml_Automation_AutomationProperties).Get(),
-                                     &automationPropertiesStatics));
-
-            THROW_IF_FAILED(automationPropertiesStatics->SetName(buttonAsDependencyObject.Get(), title.Get()));
-
-            WireButtonClickToAction(button.Get(), action, renderContext);
-        }
-
-        THROW_IF_FAILED(button.CopyTo(finalElement));
-    }
-
-    void WireButtonClickToAction(_In_ IButton* button, _In_ IAdaptiveActionElement* action, _In_ IAdaptiveRenderContext* renderContext)
-    {
-        // Note that this method currently doesn't support inline show card actions, it
-        // assumes the caller won't call this method if inline show card is specified.
-        ComPtr<IButton> localButton(button);
-        ComPtr<IAdaptiveActionInvoker> actionInvoker;
-        THROW_IF_FAILED(renderContext->get_ActionInvoker(&actionInvoker));
-        ComPtr<IAdaptiveActionElement> strongAction(action);
-
-        // Add click handler
-        ComPtr<IButtonBase> buttonBase;
-        THROW_IF_FAILED(localButton.As(&buttonBase));
-
-        EventRegistrationToken clickToken;
-        THROW_IF_FAILED(buttonBase->add_Click(Callback<IRoutedEventHandler>([strongAction, actionInvoker](IInspectable* /*sender*/, IRoutedEventArgs *
-                                                                                                          /*args*/) -> HRESULT {
-                                                  THROW_IF_FAILED(actionInvoker->SendActionEvent(strongAction.Get()));
-                                                  return S_OK;
-                                              }).Get(),
-                                              &clickToken));
     }
 
     HRESULT XamlHelpers::SetSeparatorVisibility(_In_ IPanel* parentPanel)
@@ -666,64 +554,6 @@ namespace AdaptiveNamespace::XamlHelpers
                                      &clickToken);
     }
 
-    void HandleSelectAction(_In_ IAdaptiveCardElement* adaptiveCardElement,
-                            _In_ IAdaptiveActionElement* selectAction,
-                            _In_ IAdaptiveRenderContext* renderContext,
-                            _In_ IUIElement* uiElement,
-                            bool supportsInteractivity,
-                            bool fullWidthTouchTarget,
-                            _COM_Outptr_ IUIElement** outUiElement)
-    {
-        if (selectAction != nullptr && supportsInteractivity)
-        {
-            XamlHelpers::WrapInTouchTarget(adaptiveCardElement, uiElement, selectAction, renderContext, fullWidthTouchTarget, L"Adaptive.SelectAction", outUiElement);
-        }
-        else
-        {
-            if (selectAction != nullptr)
-            {
-                renderContext->AddWarning(ABI::AdaptiveNamespace::WarningStatusCode::InteractivityNotSupported,
-                                          HStringReference(L"SelectAction present, but Interactivity is not supported").Get());
-            }
-
-            ComPtr<IUIElement> localUiElement(uiElement);
-            THROW_IF_FAILED(localUiElement.CopyTo(outUiElement));
-        }
-    }
-
-    HRESULT BuildActions(_In_ IAdaptiveCard* adaptiveCard,
-                         _In_ IVector<IAdaptiveActionElement*>* children,
-                         _In_ IPanel* bodyPanel,
-                         bool insertSeparator,
-                         _In_ IAdaptiveRenderContext* renderContext,
-                         _In_ ABI::AdaptiveNamespace::IAdaptiveRenderArgs* renderArgs)
-    {
-        ComPtr<IAdaptiveHostConfig> hostConfig;
-        RETURN_IF_FAILED(renderContext->get_HostConfig(&hostConfig));
-        ComPtr<IAdaptiveActionsConfig> actionsConfig;
-        RETURN_IF_FAILED(hostConfig->get_Actions(actionsConfig.GetAddressOf()));
-
-        // Create a separator between the body and the actions
-        if (insertSeparator)
-        {
-            ABI::AdaptiveNamespace::Spacing spacing;
-            RETURN_IF_FAILED(actionsConfig->get_Spacing(&spacing));
-
-            UINT spacingSize;
-            RETURN_IF_FAILED(GetSpacingSizeFromSpacing(hostConfig.Get(), spacing, &spacingSize));
-
-            ABI::Windows::UI::Color color = {0};
-            auto separator = XamlHelpers::CreateSeparator(renderContext, spacingSize, 0, color);
-            XamlHelpers::AppendXamlElementToPanel(separator.Get(), bodyPanel);
-        }
-
-        ComPtr<IUIElement> actionSetControl;
-        RETURN_IF_FAILED(BuildActionSetHelper(adaptiveCard, nullptr, children, renderContext, renderArgs, &actionSetControl));
-
-        XamlHelpers::AppendXamlElementToPanel(actionSetControl.Get(), bodyPanel);
-        return S_OK;
-    }
-
     HRESULT SetAutoImageSize(_In_ IFrameworkElement* imageControl, _In_ IInspectable* parentElement, _In_ IBitmapSource* imageSource, bool setVisible)
     {
         INT32 pixelHeight;
@@ -800,259 +630,5 @@ namespace AdaptiveNamespace::XamlHelpers
                 THROW_IF_FAILED(separator.CopyTo(addedSeparator));
             }
         }
-    }
-
-    HRESULT BuildActionSetHelper(_In_opt_ ABI::AdaptiveNamespace::IAdaptiveCard* adaptiveCard,
-                                 _In_opt_ IAdaptiveActionSet* adaptiveActionSet,
-                                 _In_ IVector<IAdaptiveActionElement*>* children,
-                                 _In_ IAdaptiveRenderContext* renderContext,
-                                 _In_ IAdaptiveRenderArgs* renderArgs,
-                                 _Outptr_ IUIElement** actionSetControl)
-    {
-        ComPtr<IAdaptiveHostConfig> hostConfig;
-        RETURN_IF_FAILED(renderContext->get_HostConfig(&hostConfig));
-        ComPtr<IAdaptiveActionsConfig> actionsConfig;
-        RETURN_IF_FAILED(hostConfig->get_Actions(actionsConfig.GetAddressOf()));
-
-        ABI::AdaptiveNamespace::ActionAlignment actionAlignment;
-        RETURN_IF_FAILED(actionsConfig->get_ActionAlignment(&actionAlignment));
-
-        ABI::AdaptiveNamespace::ActionsOrientation actionsOrientation;
-        RETURN_IF_FAILED(actionsConfig->get_ActionsOrientation(&actionsOrientation));
-
-        // Declare the panel that will host the buttons
-        ComPtr<IPanel> actionsPanel;
-        ComPtr<IVector<ColumnDefinition*>> columnDefinitions;
-
-        if (actionAlignment == ABI::AdaptiveNamespace::ActionAlignment::Stretch &&
-            actionsOrientation == ABI::AdaptiveNamespace::ActionsOrientation::Horizontal)
-        {
-            // If stretch alignment and orientation is horizontal, we use a grid with equal column widths to achieve
-            // stretch behavior. For vertical orientation, we'll still just use a stack panel since the concept of
-            // stretching buttons height isn't really valid, especially when the height of cards are typically dynamic.
-            ComPtr<IGrid> actionsGrid =
-                XamlHelpers::CreateXamlClass<IGrid>(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_Grid));
-            RETURN_IF_FAILED(actionsGrid->get_ColumnDefinitions(&columnDefinitions));
-            RETURN_IF_FAILED(actionsGrid.As(&actionsPanel));
-        }
-        else
-        {
-            // Create a stack panel for the action buttons
-            ComPtr<IStackPanel> actionStackPanel =
-                XamlHelpers::CreateXamlClass<IStackPanel>(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_StackPanel));
-
-            auto uiOrientation = (actionsOrientation == ABI::AdaptiveNamespace::ActionsOrientation::Horizontal) ?
-                Orientation::Orientation_Horizontal :
-                Orientation::Orientation_Vertical;
-
-            RETURN_IF_FAILED(actionStackPanel->put_Orientation(uiOrientation));
-
-            ComPtr<IFrameworkElement> actionsFrameworkElement;
-            RETURN_IF_FAILED(actionStackPanel.As(&actionsFrameworkElement));
-
-            switch (actionAlignment)
-            {
-            case ABI::AdaptiveNamespace::ActionAlignment::Center:
-                RETURN_IF_FAILED(actionsFrameworkElement->put_HorizontalAlignment(ABI::Windows::UI::Xaml::HorizontalAlignment_Center));
-                break;
-            case ABI::AdaptiveNamespace::ActionAlignment::Left:
-                RETURN_IF_FAILED(actionsFrameworkElement->put_HorizontalAlignment(ABI::Windows::UI::Xaml::HorizontalAlignment_Left));
-                break;
-            case ABI::AdaptiveNamespace::ActionAlignment::Right:
-                RETURN_IF_FAILED(actionsFrameworkElement->put_HorizontalAlignment(ABI::Windows::UI::Xaml::HorizontalAlignment_Right));
-                break;
-            case ABI::AdaptiveNamespace::ActionAlignment::Stretch:
-                RETURN_IF_FAILED(actionsFrameworkElement->put_HorizontalAlignment(ABI::Windows::UI::Xaml::HorizontalAlignment_Stretch));
-                break;
-            }
-
-            // Add the action buttons to the stack panel
-            RETURN_IF_FAILED(actionStackPanel.As(&actionsPanel));
-        }
-
-        Thickness buttonMargin;
-        RETURN_IF_FAILED(ActionHelpers::GetButtonMargin(actionsConfig.Get(), buttonMargin));
-        if (actionsOrientation == ABI::AdaptiveNamespace::ActionsOrientation::Horizontal)
-        {
-            // Negate the spacing on the sides so the left and right buttons are flush on the side.
-            // We do NOT remove the margin from the individual button itself, since that would cause
-            // the equal columns stretch behavior to not have equal columns (since the first and last
-            // button would be narrower without the same margins as its peers).
-            ComPtr<IFrameworkElement> actionsPanelAsFrameworkElement;
-            RETURN_IF_FAILED(actionsPanel.As(&actionsPanelAsFrameworkElement));
-            RETURN_IF_FAILED(actionsPanelAsFrameworkElement->put_Margin({buttonMargin.Left * -1, 0, buttonMargin.Right * -1, 0}));
-        }
-        else
-        {
-            // Negate the spacing on the top and bottom so the first and last buttons don't have extra padding
-            ComPtr<IFrameworkElement> actionsPanelAsFrameworkElement;
-            RETURN_IF_FAILED(actionsPanel.As(&actionsPanelAsFrameworkElement));
-            RETURN_IF_FAILED(actionsPanelAsFrameworkElement->put_Margin({0, buttonMargin.Top * -1, 0, buttonMargin.Bottom * -1}));
-        }
-
-        UINT32 maxActions;
-        RETURN_IF_FAILED(actionsConfig->get_MaxActions(&maxActions));
-
-        bool allActionsHaveIcons{true};
-        XamlHelpers::IterateOverVector<IAdaptiveActionElement>(children, [&](IAdaptiveActionElement* child) {
-            HString iconUrl;
-            RETURN_IF_FAILED(child->get_IconUrl(iconUrl.GetAddressOf()));
-
-            if (WindowsIsStringEmpty(iconUrl.Get()))
-            {
-                allActionsHaveIcons = false;
-            }
-            return S_OK;
-        });
-
-        UINT currentAction = 0;
-
-        RETURN_IF_FAILED(renderArgs->put_AllowAboveTitleIconPlacement(allActionsHaveIcons));
-
-        std::shared_ptr<std::vector<ComPtr<IUIElement>>> allShowCards = std::make_shared<std::vector<ComPtr<IUIElement>>>();
-        ComPtr<IStackPanel> showCardsStackPanel =
-            XamlHelpers::CreateXamlClass<IStackPanel>(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_StackPanel));
-        ComPtr<IGridStatics> gridStatics;
-        RETURN_IF_FAILED(GetActivationFactory(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_Grid).Get(), &gridStatics));
-        XamlHelpers::IterateOverVector<IAdaptiveActionElement>(children, [&](IAdaptiveActionElement* child) {
-            if (currentAction < maxActions)
-            {
-                // Render each action using the registered renderer
-                ComPtr<IAdaptiveActionElement> action(child);
-                ComPtr<IAdaptiveActionRendererRegistration> actionRegistration;
-                RETURN_IF_FAILED(renderContext->get_ActionRenderers(&actionRegistration));
-
-                ComPtr<IAdaptiveActionRenderer> renderer;
-                while (!renderer)
-                {
-                    HString actionTypeString;
-                    RETURN_IF_FAILED(action->get_ActionTypeString(actionTypeString.GetAddressOf()));
-                    RETURN_IF_FAILED(actionRegistration->Get(actionTypeString.Get(), &renderer));
-                    if (!renderer)
-                    {
-                        ABI::AdaptiveNamespace::FallbackType actionFallbackType;
-                        action->get_FallbackType(&actionFallbackType);
-                        switch (actionFallbackType)
-                        {
-                        case ABI::AdaptiveNamespace::FallbackType::Drop:
-                        {
-                            RETURN_IF_FAILED(XamlHelpers::WarnForFallbackDrop(renderContext, actionTypeString.Get()));
-                            return S_OK;
-                        }
-
-                        case ABI::AdaptiveNamespace::FallbackType::Content:
-                        {
-                            ComPtr<IAdaptiveActionElement> actionFallback;
-                            RETURN_IF_FAILED(action->get_FallbackContent(&actionFallback));
-
-                            HString fallbackTypeString;
-                            RETURN_IF_FAILED(actionFallback->get_ActionTypeString(fallbackTypeString.GetAddressOf()));
-                            RETURN_IF_FAILED(XamlHelpers::WarnForFallbackContentElement(renderContext,
-                                                                                        actionTypeString.Get(),
-                                                                                        fallbackTypeString.Get()));
-
-                            action = actionFallback;
-                            break;
-                        }
-
-                        case ABI::AdaptiveNamespace::FallbackType::None:
-                        default:
-                            return E_FAIL;
-                        }
-                    }
-                }
-
-                ComPtr<IUIElement> actionControl;
-                RETURN_IF_FAILED(renderer->Render(action.Get(), renderContext, renderArgs, &actionControl));
-
-                XamlHelpers::AppendXamlElementToPanel(actionControl.Get(), actionsPanel.Get());
-
-                ABI::AdaptiveNamespace::ActionType actionType;
-                RETURN_IF_FAILED(action->get_ActionType(&actionType));
-
-                // Build inline show cards if needed
-                if (actionType == ABI::AdaptiveNamespace::ActionType_ShowCard)
-                {
-                    ComPtr<IUIElement> uiShowCard;
-
-                    ComPtr<IAdaptiveShowCardActionConfig> showCardActionConfig;
-                    RETURN_IF_FAILED(actionsConfig->get_ShowCard(&showCardActionConfig));
-
-                    ABI::AdaptiveNamespace::ActionMode showCardActionMode;
-                    RETURN_IF_FAILED(showCardActionConfig->get_ActionMode(&showCardActionMode));
-
-                    if (showCardActionMode == ABI::AdaptiveNamespace::ActionMode::Inline)
-                    {
-                        ComPtr<IAdaptiveShowCardAction> showCardAction;
-                        RETURN_IF_FAILED(action.As(&showCardAction));
-
-                        ComPtr<IAdaptiveCard> showCard;
-                        RETURN_IF_FAILED(showCardAction->get_Card(&showCard));
-
-                        RETURN_IF_FAILED(AdaptiveShowCardActionRenderer::BuildShowCard(
-                            showCard.Get(), renderContext, renderArgs, (adaptiveActionSet == nullptr), uiShowCard.GetAddressOf()));
-
-                        ComPtr<IPanel> showCardsPanel;
-                        RETURN_IF_FAILED(showCardsStackPanel.As(&showCardsPanel));
-                        XamlHelpers::AppendXamlElementToPanel(uiShowCard.Get(), showCardsPanel.Get());
-
-                        if (adaptiveActionSet)
-                        {
-                            RETURN_IF_FAILED(
-                                renderContext->AddInlineShowCard(adaptiveActionSet, showCardAction.Get(), uiShowCard.Get()));
-                        }
-                        else
-                        {
-                            ComPtr<AdaptiveNamespace::AdaptiveRenderContext> contextImpl =
-                                PeekInnards<AdaptiveNamespace::AdaptiveRenderContext>(renderContext);
-
-                            RETURN_IF_FAILED(
-                                contextImpl->AddInlineShowCard(adaptiveCard, showCardAction.Get(), uiShowCard.Get()));
-                        }
-                    }
-                }
-
-                if (columnDefinitions != nullptr)
-                {
-                    // If using the equal width columns, we'll add a column and assign the column
-                    ComPtr<IColumnDefinition> columnDefinition = XamlHelpers::CreateXamlClass<IColumnDefinition>(
-                        HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_ColumnDefinition));
-                    RETURN_IF_FAILED(columnDefinition->put_Width({1.0, GridUnitType::GridUnitType_Star}));
-                    RETURN_IF_FAILED(columnDefinitions->Append(columnDefinition.Get()));
-
-                    ComPtr<IFrameworkElement> actionFrameworkElement;
-                    THROW_IF_FAILED(actionControl.As(&actionFrameworkElement));
-                    THROW_IF_FAILED(gridStatics->SetColumn(actionFrameworkElement.Get(), currentAction));
-                }
-            }
-            else
-            {
-                renderContext->AddWarning(ABI::AdaptiveNamespace::WarningStatusCode::MaxActionsExceeded,
-                                          HStringReference(L"Some actions were not rendered due to exceeding the maximum number of actions allowed")
-                                              .Get());
-            }
-            currentAction++;
-            return S_OK;
-        });
-
-        // Reset icon placement value
-        RETURN_IF_FAILED(renderArgs->put_AllowAboveTitleIconPlacement(false));
-
-        ComPtr<IFrameworkElement> actionsPanelAsFrameworkElement;
-        RETURN_IF_FAILED(actionsPanel.As(&actionsPanelAsFrameworkElement));
-        RETURN_IF_FAILED(XamlHelpers::SetStyleFromResourceDictionary(renderContext,
-                                                                     L"Adaptive.Actions",
-                                                                     actionsPanelAsFrameworkElement.Get()));
-
-        ComPtr<IStackPanel> actionSet =
-            XamlHelpers::CreateXamlClass<IStackPanel>(HStringReference(RuntimeClass_Windows_UI_Xaml_Controls_StackPanel));
-        ComPtr<IPanel> actionSetAsPanel;
-        actionSet.As(&actionSetAsPanel);
-
-        // Add buttons and show cards to panel
-        XamlHelpers::AppendXamlElementToPanel(actionsPanel.Get(), actionSetAsPanel.Get());
-        XamlHelpers::AppendXamlElementToPanel(showCardsStackPanel.Get(), actionSetAsPanel.Get());
-
-        return actionSetAsPanel.CopyTo(actionSetControl);
     }
 }
